updated setup steps
Ran through this with a customer and the list permission was required on the AKV storing the HSM key when they tried to create the asymmetric key or they received an error listing the 4 permissions required.  Additionally, they had to edit the registry as their SQL server service account does not have local admin access and therefore could not add the required registry key.